dom/tests/mochitest/geolocation/test_cancelCurrent.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=478911
-->
<head>
  <title>Test for getCurrentPosition </title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=478911">Mozilla Bug 478911</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

function failureCallback(error) {
  ok(error.code == error.PERMISSION_DENIED, "Ensure that the error was PERMISSION_DENIED");
  SimpleTest.finish();
}

function successCallback(position){
  ok(0, "Success was called when it shouldn't have been.  major problem");
  SimpleTest.finish();
}

/** Test for Bug  **/

SimpleTest.waitForExplicitFinish();

navigator.geolocation.getCurrentPosition(successCallback, failureCallback, null);

// click deny
setTimeout(clickNotificationButton, 50, kDenyButton);

</script>
</pre>
</body>
</html>
